+++ plugin/src/test/fixtures/junit4-plugin-kotlin-mpp-multiple-test-runs/src/commonTest/kotlin/com/example/SubTest.kt
@@ -1,0 +1,10 @@
+package com.example
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class SubTest {
+	@Test fun simple() {
+		assertEquals(1, sub(3, 2))
+	}
+}